reafct : separate and move dtos
