+++ recommend/src/main/java/com/lxp/recommend/application/port/required/CourseMetaQueryPort.java
@@ -1,6 +1,6 @@
 package com.lxp.recommend.application.port.required;
 
-import com.lxp.recommend.application.port.required.dto.CourseMetaData;
+import com.lxp.recommend.application.dto.CourseMetaData;
 
 import java.util.List;
 import java.util.Optional;
