+++ recommend/src/main/java/com/lxp/recommend/application/port/required/LearnerProfileQueryPort.java
@@ -1,6 +1,6 @@
 package com.lxp.recommend.application.port.required;
 
-import com.lxp.recommend.application.port.required.dto.LearnerProfileData;
+import com.lxp.recommend.application.dto.LearnerProfileData;
 
 import java.util.Optional;
