--- recommend/src/main/java/com/lxp/recommend/application/port/required/MemberProfileReader.java
@@ -1,8 +1,0 @@
-package com.lxp.recommend.application.port.required;
-
-import com.lxp.recommend.domain.dto.LearnerProfileView;
-
-public interface MemberProfileReader {
-    // 외부 컨텍스트 통신이므로 Long 사용 (또는 MemberId VO 사용해도 무방)
-    LearnerProfileView getProfile(String memberId);
-}
