+++ recommend/src/main/java/com/lxp/recommend/application/service/RecommendDataAssembler.java
@@ -6,9 +6,9 @@ import com.lxp.recommend.application.mapper.LearningHistoryMapper;
 import com.lxp.recommend.application.port.required.CourseMetaQueryPort;
 import com.lxp.recommend.application.port.required.LearnerProfileQueryPort;
 import com.lxp.recommend.application.port.required.LearningHistoryQueryPort;
-import com.lxp.recommend.application.port.required.dto.CourseMetaData;
-import com.lxp.recommend.application.port.required.dto.LearnerProfileData;
-import com.lxp.recommend.application.port.required.dto.LearningHistoryData;
+import com.lxp.recommend.application.dto.CourseMetaData;
+import com.lxp.recommend.application.dto.LearnerProfileData;
+import com.lxp.recommend.application.dto.LearningHistoryData;
 import com.lxp.recommend.application.service.policy.DifficultyPolicyService;
 import com.lxp.recommend.infrastructure.external.user.LearnerProfileView;
 import com.lxp.recommend.domain.model.CourseCandidate;
