+++ recommend/src/test/java/com/lxp/recommend/domain/model/RecommendContextTest.java
@@ -1,8 +1,7 @@
 package com.lxp.recommend.domain.model;
 
-import com.lxp.recommend.domain.dto.DifficultyLevel;
-import com.lxp.recommend.domain.dto.EnrollmentStatus;
-import com.lxp.recommend.domain.dto.LearningStatusView;
+import com.lxp.recommend.domain.model.ids.EnrollmentStatus;
+import com.lxp.recommend.infrastructure.external.enrollment.LearningStatusView;
 import com.lxp.recommend.domain.model.ids.CourseId;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
